Strangler extraction #15: NextSync op layer -> zxnu_nextsync_ops.py
Move the NextSync tab operation layer (~830 lines across three blocks
interleaved with SD-card ops that stay behind) out of MainWindow.__init__
as three builders called at their historical positions:

* build_nextsync_server_start - prepare-debounce timer, perform-checks/
  refresh-explorer and nextsync_start_server (Sync3/Sync4 + Remote
  Explorer -listen routing); warnings/do_server_job/_re_try_send_folder
  arrive as forwarding lambdas resolved at server-start time.
* build_nextsync_explorer_ops - the classic explorer ops: context menu,
  rename/copy/paste/import/delete, sync-root row, syncignore/syncpoint
  buttons, sync-mode + slow-transfer persistence, double-click nav.
* build_nextsync_server_job - warnings, IP info, cancel, the Sync4
  send-conflict prompt marshalling and nextsync_do_server_job.

Same seam as #10-#14: self -> host token pass, identity params for
earlier-bound closures, forwarding lambdas for later-bound ones, host
writes + call-site rebinds for the ~24 closures build_nextsync_pane and
later startup code consume by bare name (the start-server closure via
the pre-existing _nextsync_start_server_fn attr - the bare
nextsync_start_server attribute is the Start button widget).
zxnu_main.py 8,275 -> 7,513 lines. Tripwired in test_pane_imports.py.
