+++ tests/test_pane_imports.py
@@ -141,6 +141,11 @@ PANES = [
     ("zxnu_emulator_ops", "build_emulator_ops", []),
     # Extraction #13: the hdfg.cfg restore/save pipeline.
     ("zxnu_config_io", "build_config_io", []),
+    # Extraction #15: the NextSync op layer (server prepare/start, classic
+    # explorer ops, server job/warnings/conflict) — three builders.
+    ("zxnu_nextsync_ops",
+     ["build_nextsync_server_start", "build_nextsync_explorer_ops",
+      "build_nextsync_server_job"], []),
 ]
 
 for modname, funcnames, _outputs in PANES:
